Modify upgrade config to use JSON reference
