+++ configs/upgrade/strategy/TraderJoe.json
@@ -1,12 +1,14 @@
-[
-  {
-    "proxy": "0xec4042d20FCC06aBb52363f2e7f6be5396Ec79B5",
-    "newImplementation": "TraderJoe",
-    "libraries": [
-      {
-        "name": "TraderJoeInvestmentLib",
-        "dependencies": []
-      }
-    ]
-  }
-]
+{
+  "properties": [
+    {
+      "proxy": { "$ref": "../../live/strategy/TraderJoe.json#/address" },
+      "newImplementation": "TraderJoe",
+      "libraries": [
+        {
+          "name": "TraderJoeInvestmentLib",
+          "dependencies": []
+        }
+      ]
+    }
+  ]
+}
