+++ scripts/helper/paths.ts
@@ -1,7 +1,4 @@
-import { execSync } from "child_process"
-import { readFileSync, writeFileSync } from "fs"
 import path from "path"
-import { LiveConfig, UpgradeConfig } from "../interfaces/configs"
 
 const DeployConfigDir = "./configs/deploy"
 const LiveConfigDir = "./configs/live"
@@ -17,21 +14,4 @@ export function getLiveConfigPath(name: string): string {
 
 export function getUpgradeConfigPath(name: string): string {
   return path.resolve(path.join(UpgradeConfigDir, name + ".json"))
-}
-
-export function readLiveConfig(name: string): LiveConfig {
-  return JSON.parse(readFileSync(getLiveConfigPath(name), { encoding: "utf-8" }))
-}
-
-export function readUpgradeConfig(name: string): UpgradeConfig[] {
-  return JSON.parse(readFileSync(getUpgradeConfigPath(name), { encoding: "utf-8" }))
-}
-
-export function writeLiveConfig(name: string, liveConfig: LiveConfig): void {
-  const path = getLiveConfigPath(name)
-
-  writeFileSync(path, JSON.stringify(liveConfig), {
-    encoding: "utf-8",
-  })
-  execSync(`npx prettier --write '${path}'`)
 }
