+++ scripts/interfaces/configs.ts
@@ -1,5 +1,9 @@
 import { Library } from "./library"
 
+export interface DeployConfig {
+  [key: string]: any
+}
+
 export interface LiveConfig {
   name: string
   address: string
